# src/Swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "CAPTION CUỐI KHÓA"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "securityDefinitions": {},
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/comments": {
      "post": {
        "tags": [
          "Comments"
        ],
        "operationId": "newBinhLuan",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "type": "string"
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/newBinhLuanRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "get": {
        "tags": [
          "Comments"
        ],
        "operationId": "getAllBinhLuan",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/api/comments/{id}": {
      "put": {
        "tags": [
          "Comments"
        ],
        "operationId": "updateBinhLuan",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer"
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "type": "string"
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/updateBinhLuanRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "delete": {
        "tags": [
          "Comments"
        ],
        "operationId": "deleteBinhLuan",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer"
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/api/comments/lay-binh-luan-theo-phong/{id}": {
      "get": {
        "tags": [
          "Comments"
        ],
        "operationId": "getBinhLuanByRoom",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer"
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/api/booking": {
      "post": {
        "tags": [
          "Booking"
        ],
        "operationId": "newBooking",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "type": "string"
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/newBookingRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "get": {
        "tags": [
          "Booking"
        ],
        "operationId": "getAllBooking",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/api/booking/{id}": {
      "get": {
        "tags": [
          "Booking"
        ],
        "operationId": "getBookingById",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer"
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "put": {
        "tags": [
          "Booking"
        ],
        "operationId": "updateBookingData",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer"
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "type": "string"
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/updateBookingDataRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "delete": {
        "tags": [
          "Booking"
        ],
        "operationId": "deleteBooking",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer"
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/api/booking/lay-theo-nguoi-dung/{ma_nguoi_dat}": {
      "get": {
        "tags": [
          "Booking"
        ],
        "operationId": "getBookingByUserId",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "ma_nguoi_dat",
            "in": "path",
            "required": true,
            "type": "integer"
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "newBinhLuanRequest": {
      "title": "newBinhLuanRequest",
      "example": {
        "ma_phong": 5,
        "ma_nguoi_binh_luan": 3,
        "noi_dung": "This is a comment",
        "sao_binh_luan": 5
      },
      "type": "object",
      "properties": {
        "ma_phong": {
          "type": "integer",
          "format": "int32"
        },
        "ma_nguoi_binh_luan": {
          "type": "integer",
          "format": "int32"
        },
        "noi_dung": {
          "type": "string"
        },
        "sao_binh_luan": {
          "type": "integer",
          "format": "int32"
        }
      },
      "required": [
        "ma_phong",
        "ma_nguoi_binh_luan",
        "noi_dung",
        "sao_binh_luan"
      ]
    },
    "updateBinhLuanRequest": {
      "title": "updateBinhLuanRequest",
      "example": {
        "noi_dung": "This is a commentz",
        "sao_binh_luan": 5
      },
      "type": "object",
      "properties": {
        "noi_dung": {
          "type": "string"
        },
        "sao_binh_luan": {
          "type": "integer",
          "format": "int32"
        }
      },
      "required": [
        "noi_dung",
        "sao_binh_luan"
      ]
    },
    "newBookingRequest": {
      "title": "newBookingRequest",
      "example": {
        "ma_phong": 5,
        "ngay_den": "00-00-0000",
        "ngay_di": "00-00-0000",
        "so_luong_khach": 4
      },
      "type": "object",
      "properties": {
        "ma_phong": {
          "type": "integer",
          "format": "int32"
        },
        "ngay_den": {
          "type": "string"
        },
        "ngay_di": {
          "type": "string"
        },
        "so_luong_khach": {
          "type": "integer",
          "format": "int32"
        }
      },
      "required": [
        "ma_phong",
        "ngay_den",
        "ngay_di",
        "so_luong_khach"
      ]
    },
    "updateBookingDataRequest": {
      "title": "updateBookingDataRequest",
      "example": {
        "ma_phong": 5,
        "ngay_den": "00-00-0000",
        "ngay_di": "00-00-0000",
        "so_luong_khach": 2
      },
      "type": "object",
      "properties": {
        "ma_phong": {
          "type": "integer",
          "format": "int32"
        },
        "ngay_den": {
          "type": "string"
        },
        "ngay_di": {
          "type": "string"
        },
        "so_luong_khach": {
          "type": "integer",
          "format": "int32"
        }
      },
      "required": [
        "ma_phong",
        "ngay_den",
        "ngay_di",
        "so_luong_khach"
      ]
    },
...
  },
  "security": [],
...
}